Effects: cast EQ output in place (drop per-block mixer-thread allocation)
Adversarial review flagged EQProcessor.process allocating a full-size float32
temporary via astype every block on the real-time mixer thread. block[:] = y
performs the float64->float32 downcast directly into the destination buffer —
identical result, one fewer per-block heap allocation.

+++ src/subtitld/modules/audio_effects.py
@@ -156,7 +156,9 @@ class EQProcessor:
         if self.sos is None or _sig is None or block.shape[0] == 0:
             return
         y, self.zi = _sig.sosfilt(self.sos, block, axis=0, zi=self.zi)
-        block[:] = y.astype(block.dtype, copy=False)
+        # Cast float64 → block's float32 directly into the destination buffer
+        # (no full-size temporary — this runs on the real-time mixer thread).
+        block[:] = y
 
 
 # --------------------------------------------------------------------------- #
